requirements_local.txt
absl-py==0.6.1
aiohttp==3.4.4
aiosqlite==0.8.0
appnope==0.1.0
astor==0.7.1
astroid==2.1.0
async-timeout==3.0.1
attrs==18.2.0
backcall==0.1.0
bleach==3.0.2
certifi==2018.11.29
chardet==3.0.4
decorator==4.3.0
defusedxml==0.5.0
entrypoints==0.2.3
gast==0.2.0
grpcio==1.17.0
h5py==2.8.0
idna==2.8
idna-ssl==1.1.0
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
jedi==0.13.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
lazy-object-proxy==1.3.1
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
multidict==4.5.2
nbconvert==5.4.0
nbformat==4.4.0
nltk==3.4
notebook==5.7.2
numpy==1.15.4
opencv-python==3.4.4.19
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pathlib==1.0.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.3.0
pylint==2.2.2
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
scipy==1.1.0
Send2Trash==1.5.0
singledispatch==3.4.0.3
six==1.11.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.1
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.10.11
yarl==1.3.0
